Refactor: 세션 이미지 관련 로직를 세션 서비스로부터 분리 (#81)
